refactor: Move contentserver implementation into views.py
Minimal adjustment of affected import paths, no lint cleanup.

This is the second commit in a refactor/change series.
